restructuring of directory tree
